requirements.txt
# Core Dependencies
streamlit>=1.31.0
streamlit-extras>=0.3.6

# AI/LLM Framework
openai>=1.12.0
anthropic>=0.18.1

# Vector Store & RAG
faiss-cpu>=1.9.0
sentence-transformers>=3.3.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.6.0

# MCP Server
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
colorlog==6.8.2
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
